requirements.txt
# Core
fastapi>=0.109.0
uvicorn>=0.27.0
pydantic>=2.5.0
pydantic-settings>=2.1.0

# Snowflake
snowflake-connector-python>=3.6.0
snowflake-snowpark-python>=1.11.0

# LLM
anthropic>=0.18.0

# Frontend
streamlit>=1.30.0
plotly>=5.18.0

# Utilities
python-dotenv>=1.0.0
pandas>=2.1.0
numpy>=1.26.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.23.0
httpx>=0.26.0
